[Refactor] 네이버 API 응답 수정
